Made it possible for generators to process symlinks.
git-svn-id: https://svn.code.sf.net/p/tsc-tpm/code@381 af914a7b-af0d-0410-90d3-cea0664cafdb

+++ distribution/installation/generator.rb
@@ -97,12 +97,12 @@ module Installation
     end
 
     def process_create
-      result = stream_or_nothing(target, saved_target) do |_input|
+      result = stream_or_nothing(target, saved_target) { |_input|
         create _input
-      end
+      }
+      return unless result
 
       FileUtils.remove_entry target rescue true
-
       File.open(target, 'w') do |_io|
         _io.puts result
       end
